Reschedule jobs #39 (refactoring)
